requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
streamlit==1.25.0
nltk==3.8.1
requests==2.31.0
tmdbv3api==1.9.0
python-dotenv==1.0.0